chore!: restart the project with a better design
The next design solves the difference problems between pawns and every
other piece by making it its own, distinct class.

The algorithm to deal with pawns is different, so now board will be
injected into a piece/pawn and more of the algorithm is implemented at
the piece/pawn level.

I will also change the moveset from a list of all moves disregarding
limitations from the board instead to base moves that can be used to
build the other moves up to a given limit (or no limit). Now, each move
is validated as it is created in a loop, which stops at the limit or
when the piece is blocked from moving further (either by a piece or the
edge of the board).

The responsibilities will be divided as follows, subject to change:
- Piece/Pawn will return the list of moves ignoring King safety
- Board will collect all piece moves together for a chosen color, also
  adding the special moves En Passant and Castling as needed
- Game will go through each move and create a separate Board object to
  generate movesets as needed to check for own King safety as well as
  checking for check, stalemate, and checkmate. Castling moves have
  extra restrictions which are checked here too. Finally, the list of
  all valid moves for a given color for the current Board state is
  returned

These objects will have other responsibilities but they should be much
simpler to implement... Otherwise, the design is the same as before.

BREAKING CHANGE: Nothing from before will work!